requirements.txt
asgiref==3.5.0
Django==4.0.1
gunicorn==20.1.0
pypng==0.0.21
PyQRCode==1.2.1
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
